test: allow short benchmarks for tests
Enable short benchmarks for process and path benchmark tests. These have
been observed to fail in CI due to returning with zero operations
performed in the allotted time. We have a special environment variable
for other benchmark tests that can be set to make that allowable in
benchmarks. Set it for path and process.

PR-URL: https://github.com/nodejs/node/pull/16097
Reviewed-By: Evan Lucas <evanlucas@me.com>
Reviewed-By: Minwoo Jung <minwoo@nodesource.com>
Reviewed-By: Gibson Fahnestock <gibfahn@gmail.com>